This is about as far as we can get for RS5 in terms of detangling the mess between `COOKED_READ_DATA`, `CLE_POPUP`, and Command Line functionality (all of which use `_output.cpp`/`_output.h` functions to draw their overlays/popups/editing lines into the buffer and erase them later.)

The major workstreams were as follows:
- Popup-related materials from `CLE_POPUP` have been converted into the `Popup` class. All methods that invoked primarily on a Popup pointer have joined the class. Many of the variables have been hidden. Many methods internal to just drawing and managing popup state, its backups of `SCREEN_INFORMATION` and the way it instantiates and cleans up are moved into `Popup`.
- Started to class-ify `COOKED_READ_DATA` by moving construction into a constructor, moving methods onto the class, and beginning to hide parameters. Also very importantly, started to hand out `gsl::span` referring to the assorted pointers maintaining the internal state.
- Very beginnings of added some Command line methods to the `CommandLine` class. This may not survive long term as many of them probably belong to `COOKED_READ_DATA` or `Popup`, but it's a reasonable interim to start peeling apart this mess.
- History information moved into `CommandHistory` class. Manages its own memory. Stores information as `std::wstring`. No more tampering with the internals of history lists. Static list moved from the giant `CONSOLE_INFORMATION` global onto a static instance on this class.
- Replaced some `memmove`s with safer operations (like `std::copy` or `std::copy_n`)
- Favored `size_t` over smaller data types
- Used `gsl::narrow` in many cases to couple `size_t` to `SHORT` (some C-casts remain)
- Minorly reduced complexity of cancellation off of `WaitRoutine`
- Moved code toward references where possible
- No more 0 length arrays. Warning suppression can be removed.
- No more DDK linked lists. We can remove those headers.
- Torched a ton of macros. Converted to methods where necessary and removed the rest.
- Removed a ton of OACR suppressions about buffer overflows by risky behaviors to `std::wstring`, `std::wstring_view`, or `gsl::span`.
- There's likely a handful of other things in here too.

Related work items: #10211332, #16115769, #614402, #9765778

+++ src/host/misc.h
@@ -28,12 +28,12 @@ WCHAR CharToWchar(_In_reads_(cch) const char * const pch, const UINT cch);
 void SetConsoleCPInfo(const BOOL fOutput);
 
 BOOL CheckBisectStringW(_In_reads_bytes_(cBytes) const WCHAR * pwchBuffer,
-                        _In_ DWORD cWords,
-                        _In_ DWORD cBytes);
+                        _In_ size_t cWords,
+                        _In_ size_t cBytes);
 BOOL CheckBisectProcessW(const SCREEN_INFORMATION& ScreenInfo,
                          _In_reads_bytes_(cBytes) const WCHAR * pwchBuffer,
-                         _In_ DWORD cWords,
-                         _In_ DWORD cBytes,
+                         _In_ size_t cWords,
+                         _In_ size_t cBytes,
                          _In_ SHORT sOriginalXPosition,
                          _In_ BOOL fEcho);
